pyproject.toml
[tool.poetry]
name = "drukarnia-api"
version = "0.1.0"
description = ""
authors = [
    "Psychotim",
    "androu"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "3.9.0b0"
attrdict = "2.0.1"
fake-useragent = "1.1.3"
attrs = "23.2.0"
typeguard = "^4.1.5"

[tool.poetry.dev-dependencies]
mypy = "^1.8.0"
ruff = "^0.2.2"

[tool.mypy]
strict = true
show_error_codes = true
show_error_context = true
pretty = true
ignore_missing_imports = false
warn_unused_configs = true
disallow_subclassing_any = false
disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
follow_imports_for_stubs = true
namespace_packages = true
show_absolute_path = true
disable_error_code = []

[tool.ruff]
fix = true
unsafe-fixes = true
line-length = 120  # if active, remove E501 from ignore
lint.select = ["ALL"]
lint.ignore = [
    "RUF009",    # Do not perform function call `Join` in dataclass defaults
    "D1", "D203", "D213", "FA102", "ANN101", "UP007", "FA100", "COM812", "ISC001", "ANN401"
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
